Replaced UserAdmin based admin from with ModelForm
