+++ profiles/forms.py
@@ -1,10 +1,9 @@
 from django import forms
 from django.conf import settings
 from django.db.models import get_model
-from django.forms.util import ErrorList
 from django.utils.translation import ugettext_lazy as _
 
-from profiles.models import Profile
+from profiles.models import Profile, BasePasswordChangeForm
 
 
 try:
@@ -12,36 +11,9 @@ try:
 except AttributeError:
     from django.contrib.auth.models import User as user_model
 
-class UserChangeForm(forms.ModelForm):
-    password1 = forms.CharField(label=_('Password'), required=False,
-                                widget=forms.PasswordInput)
-    password2 = forms.CharField(label=_('Password confirmation'),
-                                required=False, widget=forms.PasswordInput)
-
+class UserChangeForm(BasePasswordChangeForm):
     class Meta:
         model = user_model
-
-    def clean(self):
-        cleaned_data = self.cleaned_data
-        password1 = cleaned_data.get('password1', '')
-        password2 = cleaned_data.get('password2', '')
-        if password1 != password2:
-            msg = u'The password fields did not match.'
-            self._errors['password2'] = ErrorList([msg])
-            del cleaned_data['password1']
-            del cleaned_data['password2']
-
-        return cleaned_data
-
-    def save(self, commit=True):
-        user = super(UserChangeForm, self).save(commit=False)
-        if self.cleaned_data['password1'] and self.cleaned_data['password2']:
-            user.set_password(self.cleaned_data['password1'])
-        if commit:
-            user.save()
-            self.save_m2m()
-        return user
-
 
 class ProfileForm(UserChangeForm):
     class Meta:
